requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
websockets==12.0
python-multipart==0.0.6
ultralytics==8.0.196
numpy==1.24.3
opencv-python==4.8.1.78
python-jose==3.3.0
python-dotenv==1.0.0
pydantic==2.1.1
pillow==10.0.1
sqlalchemy==2.0.19
psycopg2-binary==2.9.6
passlib==1.7.4
bcrypt==4.0.1
alembic==1.11.1 